fix: token count error (#6680)
